requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
pydantic==2.9.2
python-dotenv==1.0.1
httpx==0.27.2
beautifulsoup4==4.12.3
langgraph==0.2.39
langchain-core==0.3.13
langchain==0.3.7
google-generativeai==0.8.3
pinecone-client==5.0.1
serpapi==0.1.5
streamlit==1.39.0
pydantic-settings==2.6.1
orjson==3.10.7
sentence-transformers==2.2.2
transformers==4.33.3
tokenizers==0.13.3
huggingface-hub==0.17.3
pypdf==3.17.0
